% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppclp.R
\name{ppclp2D}
\alias{ppclp2D}
\title{2D prbabilistic principal curve with length penalty}
\usage{
ppclp2D(x, y, x_fix, y_fix, K = 50, degree_free = 10, lambda = 0.2,
  T = 100)
}
\arguments{
\item{x}{The coordinate of the x axis}

\item{y}{The coordinate of the y axis}

\item{x_fix}{The cordinate of the starting point of the curve}

\item{y_fix}{The coordinate of the ending point of the curve}

\item{K}{the number of knots of the curve}

\item{degree_free}{the degree of freedom of the B spline}

\item{lambda}{The magnitude value added to the length penalty}

\item{T}{The total number of iterations to do the EM algorithm}
}
\description{
This function applies the probabilistic principal curve algorithm with length penalty
}
